fastapi-backend/requirements.txt
# Core Framework
fastapi==0.110.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6

# Async HTTP & WebSocket
httpx==0.25.2
aiohttp==3.9.1
websockets==12.0

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.0
redis[hiredis]==5.0.1

# Task Queue
celery[redis]==5.3.4
flower==2.0.1
dramatiq[redis,watch]==1.15.0

# Data Validation & Serialization
pydantic==1.10.13
pydantic[email]==1.10.13
email-validator==2.1.0
orjson==3.9.10

# Security
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
cryptography==41.0.7

# OSINT Tools
python-whois==0.8.0
dnspython==2.4.2
shodan==1.31.0
tweepy==4.14.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0
python-json-logger==2.0.7
sentry-sdk==1.39.1
psutil==5.9.8

# Utils
tenacity==8.2.3
pytz==2023.3
python-dateutil==2.8.2
aiofiles==23.2.1

# Email
aiosmtplib==3.0.1
jinja2==3.1.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0